pyproject.toml
[project]
name = "conversation-data-generate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.28.0",
    "loguru>=0.7.3",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
